Add pinned behavior
